+++ src/test/java/Unicorn_00/CategoryTest.java
@@ -1,6 +1,5 @@
 package Unicorn_00;
 import org.junit.jupiter.api.Test;
-import java.math.BigDecimal;
 import static org.junit.jupiter.api.Assertions.*;
 
 public class CategoryTest {
@@ -8,7 +7,7 @@ public class CategoryTest {
     @Test
     public void testAddProduct() {
         Category category = new Category();
-        Product product = new Product(1, "Test Product", BigDecimal.valueOf(10.00), "Test Description", 5, "test.jpg");
+        Product product = new Product(1, "Test Product", 10.00, "Test Description", 5, "test.jpg");
         category.addProduct(product);
         assertTrue(category.getProducts().contains(product));
     }
@@ -16,7 +15,7 @@ public class CategoryTest {
     @Test
     public void testRemoveProduct() {
         Category category = new Category();
-        Product product = new Product(1, "Test Product", BigDecimal.valueOf(10.00), "Test Description", 5, "test.jpg");
+        Product product = new Product(1, "Test Product", 10.00, "Test Description", 5, "test.jpg");
         category.addProduct(product);
         category.removeProduct(product);
         assertFalse(category.getProducts().contains(product));
