+++ src/test/java/Unicorn_00/ProductTest.java
@@ -7,14 +7,14 @@ public class ProductTest {
 
     @Test
     public void testUpdateStock() {
-        Product product = new Product(1, "Test Product", BigDecimal.valueOf(10.00), "Test Description", 5, "test.jpg");
+        Product product = new Product(1, "Test Product", 10.00, "Test Description", 5, "test.jpg");
         product.updateStock(3);
         assertEquals(8, product.getStockQuantity());
     }
 
     @Test
     public void testAddCategory() {
-        Product product = new Product(1, "Test Product", BigDecimal.valueOf(10.00), "Test Description", 5, "test.jpg");
+        Product product = new Product(1, "Test Product", 10.00, "Test Description", 5, "test.jpg");
         Category category = new Category();
         product.addCategory(category);
         assertTrue(product.getCategories().contains(category));
@@ -22,7 +22,7 @@ public class ProductTest {
 
     @Test
     public void testRemoveCategory() {
-        Product product = new Product(1, "Test Product", BigDecimal.valueOf(10.00), "Test Description", 5, "test.jpg");
+        Product product = new Product(1, "Test Product", 10.00, "Test Description", 5, "test.jpg");
         Category category = new Category();
         product.addCategory(category);
         product.removeCategory(category);
